MR #7: Task Tool Package
